Restart by BroadcastReceiver.

+++ patch-util/src/main/java/cn/mutils/app/patch/KillAppThread.java
@@ -1,23 +1,19 @@
 package cn.mutils.app.patch;
 
-import android.app.AlarmManager;
-import android.app.PendingIntent;
-import android.content.Context;
-import android.content.Intent;
 import android.os.Process;
 import android.util.Log;
 
 /**
  * Created by wenhua.ywh on 2016/12/9.
  */
-public class RestartAppThread extends Thread {
+public class KillAppThread extends Thread {
 
-    private Context mContext;
+    private SoHotfix mSoHotfix;
     private int mDelaySeconds;
 
-    public RestartAppThread(Context context, int delaySeconds) {
+    public KillAppThread(SoHotfix soHotfix, int delaySeconds) {
         super();
-        mContext = context;
+        mSoHotfix = soHotfix;
         mDelaySeconds = delaySeconds;
     }
 
@@ -26,13 +22,13 @@ public class RestartAppThread extends Thread {
         while (true) {
             boolean isAlwaysBackground = true;
             for (int i = 0; i < mDelaySeconds; i++) {
-                Log.d(RestartAppThread.class.getSimpleName(), "run:" + i);
+                Log.d(KillAppThread.class.getSimpleName(), "run " + i);
                 try {
                     Thread.sleep(1000L);
                 } catch (Exception e) {
                     e.printStackTrace();
                 }
-                if (!SoHotfixUtil.isAppInBackground(mContext)) {
+                if (!SoHotfixUtil.isAppInBackground(mSoHotfix.getSoContext().getContext()) && !mSoHotfix.isAppWorking()) {
                     isAlwaysBackground = false;
                     break;
                 }
@@ -45,10 +41,9 @@ public class RestartAppThread extends Thread {
     }
 
     private void restartApp() {
-        Intent intent = mContext.getPackageManager().getLaunchIntentForPackage(mContext.getPackageName());
-        PendingIntent restartIntent = PendingIntent.getActivity(mContext, 0, intent, PendingIntent.FLAG_ONE_SHOT);
-        AlarmManager mgr = (AlarmManager) mContext.getSystemService(Context.ALARM_SERVICE);
-        mgr.set(AlarmManager.RTC, System.currentTimeMillis() + 500, restartIntent);
+        if (mSoHotfix.onInterceptKillApp()) {
+            return;
+        }
         System.exit(0);
         Process.killProcess(Process.myPid());
     }
